Moves alerts to Splunk package and removes redundant alerts pkg
The alerts package contained only the alert struct at this point, so the
alert was redefined as splunk.Alert, and the now-redundant alert package
removed.
